refactor: reorganize CCS module layout
