+++ src/ccs/index.ts
@@ -1,0 +1,2 @@
+export { SourceControlApi } from "./sourcecontrol/client";
+export { resolveContextExpression } from "./commands/contextHelp";
